Make changes and minor fixes to test script for Archiving list request
Fix test script for Archiving list, where the test failed when the response is an empty array.
Change status for cases when we don't have a recording for some date range. Instead of the previous marked Failed status, now its a PASS with a similar msg that doesn't have any recordings for the given Date range.
Added checks for new fields that are added to the response body, CDR and Chats.